+++ Frameworks/Foundation.framework/NSConditionLock.h
@@ -8,21 +8,21 @@
 
 + (id)allocWithZone:(struct _NSZone { }*)arg1;
 
-- (int)condition;
+- (NSInteger)condition;
 - (void)dealloc;
 - (id)description;
 - (void)finalize;
 - (id)init;
-- (id)initWithCondition:(int)arg1;
+- (id)initWithCondition:(NSInteger)arg1;
 - (void)lock;
 - (BOOL)lockBeforeDate:(id)arg1;
-- (BOOL)lockWhenCondition:(int)arg1 beforeDate:(id)arg2;
-- (void)lockWhenCondition:(int)arg1;
+- (BOOL)lockWhenCondition:(NSInteger)arg1 beforeDate:(id)arg2;
+- (void)lockWhenCondition:(NSInteger)arg1;
 - (id)name;
 - (void)setName:(id)arg1;
 - (BOOL)tryLock;
-- (BOOL)tryLockWhenCondition:(int)arg1;
+- (BOOL)tryLockWhenCondition:(NSInteger)arg1;
 - (void)unlock;
-- (void)unlockWithCondition:(int)arg1;
+- (void)unlockWithCondition:(NSInteger)arg1;
 
 @end
