+++ Frameworks/Foundation.framework/NSDebugString.h
@@ -8,12 +8,12 @@
     NSString *string;
 }
 
-- (unsigned short)characterAtIndex:(unsigned int)arg1;
+- (unsigned short)characterAtIndex:(NSUInteger)arg1;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (void)dealloc;
 - (void)finalize;
-- (void)getCharacters:(unsigned short*)arg1 range:(struct _NSRange { unsigned int x1; unsigned int x2; })arg2;
+- (void)getCharacters:(unsigned short*)arg1 range:(struct _NSRange { NSUInteger x1; NSUInteger x2; })arg2;
 - (id)initWithString:(id)arg1;
-- (unsigned int)length;
+- (NSUInteger)length;
 
 @end
