+++ Frameworks/Foundation.framework/NSKeyValueFastMutableArray.h
@@ -11,12 +11,12 @@
 - (id)_proxyInitWithContainer:(id)arg1 getter:(id)arg2;
 - (void)_proxyNonGCFinalize;
 - (void)addObject:(id)arg1;
-- (void)insertObject:(id)arg1 atIndex:(unsigned int)arg2;
+- (void)insertObject:(id)arg1 atIndex:(NSUInteger)arg2;
 - (void)insertObjects:(id)arg1 atIndexes:(id)arg2;
 - (void)removeLastObject;
-- (void)removeObjectAtIndex:(unsigned int)arg1;
+- (void)removeObjectAtIndex:(NSUInteger)arg1;
 - (void)removeObjectsAtIndexes:(id)arg1;
-- (void)replaceObjectAtIndex:(unsigned int)arg1 withObject:(id)arg2;
+- (void)replaceObjectAtIndex:(NSUInteger)arg1 withObject:(id)arg2;
 - (void)replaceObjectsAtIndexes:(id)arg1 withObjects:(id)arg2;
 
 @end
