+++ Frameworks/Foundation.framework/NSKeyValueSlowMutableArray.h
@@ -11,7 +11,7 @@
     NSKeyValueSetter *_valueSetter;
 }
 
-+ (struct { unsigned int x1; id x2[4]; }*)_proxyNonGCPoolPointer;
++ (struct { NSUInteger x1; id x2[4]; }*)_proxyNonGCPoolPointer;
 
 - (id)_createNonNilMutableArrayValueWithSelector:(SEL)arg1;
 - (id)_nonNilArrayValueWithSelector:(SEL)arg1;
@@ -19,16 +19,16 @@
 - (void)_proxyNonGCFinalize;
 - (void)_raiseNilValueExceptionWithSelector:(SEL)arg1;
 - (void)addObject:(id)arg1;
-- (unsigned int)count;
-- (void)getObjects:(id*)arg1 range:(struct _NSRange { unsigned int x1; unsigned int x2; })arg2;
-- (void)insertObject:(id)arg1 atIndex:(unsigned int)arg2;
+- (NSUInteger)count;
+- (void)getObjects:(id*)arg1 range:(struct _NSRange { NSUInteger x1; NSUInteger x2; })arg2;
+- (void)insertObject:(id)arg1 atIndex:(NSUInteger)arg2;
 - (void)insertObjects:(id)arg1 atIndexes:(id)arg2;
-- (id)objectAtIndex:(unsigned int)arg1;
+- (id)objectAtIndex:(NSUInteger)arg1;
 - (id)objectsAtIndexes:(id)arg1;
 - (void)removeLastObject;
-- (void)removeObjectAtIndex:(unsigned int)arg1;
+- (void)removeObjectAtIndex:(NSUInteger)arg1;
 - (void)removeObjectsAtIndexes:(id)arg1;
-- (void)replaceObjectAtIndex:(unsigned int)arg1 withObject:(id)arg2;
+- (void)replaceObjectAtIndex:(NSUInteger)arg1 withObject:(id)arg2;
 - (void)replaceObjectsAtIndexes:(id)arg1 withObjects:(id)arg2;
 
 @end
