+++ Frameworks/Foundation.framework/NSPathStore2.h
@@ -4,22 +4,22 @@
 
 @interface NSPathStore2 : NSString {
     unsigned short _characters[0];
-    unsigned int _lengthAndRefCount;
+    NSUInteger _lengthAndRefCount;
 }
 
-+ (id)pathStoreWithCharacters:(const unsigned short*)arg1 length:(unsigned int)arg2;
++ (id)pathStoreWithCharacters:(const unsigned short*)arg1 length:(NSUInteger)arg2;
 + (id)pathWithComponents:(id)arg1;
 
 - (id)_stringByResolvingSymlinksInPathUsingCache:(BOOL)arg1;
 - (id)_stringByStandardizingPathUsingCache:(BOOL)arg1;
-- (unsigned short)characterAtIndex:(unsigned int)arg1;
+- (unsigned short)characterAtIndex:(NSUInteger)arg1;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
-- (void)getCharacters:(unsigned short*)arg1 range:(struct _NSRange { unsigned int x1; unsigned int x2; })arg2;
-- (unsigned int)hash;
+- (void)getCharacters:(unsigned short*)arg1 range:(struct _NSRange { NSUInteger x1; NSUInteger x2; })arg2;
+- (NSUInteger)hash;
 - (BOOL)isAbsolutePath;
 - (BOOL)isEqualToString:(id)arg1;
 - (id)lastPathComponent;
-- (unsigned int)length;
+- (NSUInteger)length;
 - (id)pathComponents;
 - (id)pathExtension;
 - (id)stringByAbbreviatingWithTildeInPath;
