+++ Frameworks/Foundation.framework/NSURLHostNameAddressInfo.h
@@ -3,16 +3,16 @@
  */
 
 @interface NSURLHostNameAddressInfo : NSObject {
-    struct addrinfo { int x1; int x2; int x3; int x4; unsigned int x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; } *addressInfo;
+    struct addrinfo { NSInteger x1; NSInteger x2; NSInteger x3; NSInteger x4; NSUInteger x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; } *addressInfo;
     double timestamp;
 }
 
 + (id)addressInfoForHost:(id)arg1;
 + (void)asyncResolveWithCallbackClient:(id)arg1;
 
-- (id)_initWithAddressInfo:(struct addrinfo { int x1; int x2; int x3; int x4; unsigned int x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; }*)arg1;
+- (id)_initWithAddressInfo:(struct addrinfo { NSInteger x1; NSInteger x2; NSInteger x3; NSInteger x4; NSUInteger x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; }*)arg1;
 - (double)_timestamp;
-- (struct addrinfo { int x1; int x2; int x3; int x4; unsigned int x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; }*)addrinfo;
+- (struct addrinfo { NSInteger x1; NSInteger x2; NSInteger x3; NSInteger x4; NSUInteger x5; char *x6; struct sockaddr {} *x7; struct addrinfo {} *x8; }*)addrinfo;
 - (void)dealloc;
 - (void)finalize;
