+++ Frameworks/Foundation.framework/_NSKeyedCoderOldStyleArray.h
@@ -4,18 +4,18 @@
 
 @interface _NSKeyedCoderOldStyleArray : NSObject {
     void *_addr;
-    unsigned int _count;
+    NSUInteger _count;
     BOOL _decoded;
     BOOL _padding[2];
-    unsigned int _size;
+    NSUInteger _size;
     BOOL _type;
 }
 
 - (void)dealloc;
 - (void)encodeWithCoder:(id)arg1;
-- (void)fillObjCType:(BOOL)arg1 count:(unsigned int)arg2 at:(void*)arg3;
+- (void)fillObjCType:(BOOL)arg1 count:(NSUInteger)arg2 at:(void*)arg3;
 - (void)finalize;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithObjCType:(BOOL)arg1 count:(unsigned int)arg2 at:(const void*)arg3;
+- (id)initWithObjCType:(BOOL)arg1 count:(NSUInteger)arg2 at:(const void*)arg3;
 
 @end
