+++ Frameworks/UIKit.framework/UIAutocorrectShadowView.h
@@ -3,10 +3,10 @@
  */
 
 @interface UIAutocorrectShadowView : UIView {
-    int m_edgeType;
+    NSInteger m_edgeType;
 }
 
 - (void)drawRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
-- (void)setEdgeType:(int)arg1;
+- (void)setEdgeType:(NSInteger)arg1;
 
 @end
