+++ Frameworks/UIKit.framework/UICompositeImageView.h
@@ -8,8 +8,8 @@
     NSMutableArray *m_images;
 }
 
-- (void)addImage:(id)arg1 operation:(int)arg2 fraction:(float)arg3;
-- (void)addImage:(id)arg1 toRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg2 fromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg3 operation:(int)arg4 fraction:(float)arg5;
+- (void)addImage:(id)arg1 operation:(NSInteger)arg2 fraction:(float)arg3;
+- (void)addImage:(id)arg1 toRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg2 fromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg3 operation:(NSInteger)arg4 fraction:(float)arg5;
 - (void)addImage:(id)arg1 toRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg2 fromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg3;
 - (void)addImage:(id)arg1;
 - (void)dealloc;
