+++ Frameworks/UIKit.framework/UIGestureAnimation.h
@@ -3,12 +3,12 @@
  */
 
 @interface UIGestureAnimation : UIAnimation {
-    int _gestureType;
+    NSInteger _gestureType;
     SEL _progressSelector;
 }
 
-- (int)gestureType;
-- (id)initWithTarget:(id)arg1 progressSelector:(SEL)arg2 gestureType:(int)arg3;
+- (NSInteger)gestureType;
+- (id)initWithTarget:(id)arg1 progressSelector:(SEL)arg2 gestureType:(NSInteger)arg3;
 - (void)setProgress:(float)arg1;
 
 @end
